+++ tests/V1/Assurances/EnsureAnyOneOfTest.php
@@ -45,6 +45,8 @@ namespace GanbaroDigitalTest\Defensive\V1\Assurances;
 
 use GanbaroDigital\Defensive\V1\Exceptions\UnsupportedType;
 use GanbaroDigital\Defensive\V1\Assurances\EnsureAnyOneOf;
+use GanbaroDigital\Defensive\V1\Assurances\InvokeableAssurance;
+use GanbaroDigital\Defensive\V1\Assurances\ListableAssurance;
 use GanbaroDigital\Defensive\V1\Interfaces\Assurance;
 use PHPUnit_Framework_TestCase;
 use stdClass;
@@ -242,6 +244,55 @@ class EnsureAnyOneOfTest extends PHPUnit_Framework_TestCase
         EnsureAnyOneOf::apply($assurances)->to($item, "\$item");
     }
 
+    /**
+     * @covers ::apply
+     * @covers ::toList
+     */
+    public function test_can_apply_to_a_data_list()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $assurances = [
+            new EnsureAnyOneOfTest_EnsureString,
+            new EnsureAnyOneOfTest_EnsureNumeric,
+        ];
+
+        $list = [
+            "0",
+            "1.0",
+            "100"
+        ];
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        // if these do not match, an exception is thrown
+        EnsureAnyOneOf::apply($assurances)->toList($list, 'value');
+    }
+
+    /**
+     * @covers ::apply
+     * @covers ::toList
+     * @dataProvider provideNonLists
+     * @expectedException InvalidArgumentException
+     */
+    public function test_throws_InvalidArgumentException_if_non_list_passed_to_toList($list)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $assurances = [
+            new EnsureAnyOneOfTest_EnsureString,
+            new EnsureAnyOneOfTest_EnsureNumeric,
+        ];
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        EnsureAnyOneOf::apply($assurances)->toList($list);
+    }
+
     public function provideBadAssurances()
     {
         return [
@@ -283,14 +334,25 @@ class EnsureAnyOneOfTest extends PHPUnit_Framework_TestCase
             [ new stdClass ]
         ];
     }
+
+    public function provideNonLists()
+    {
+        return [
+            [ null ],
+            [ false ],
+            [ true ],
+            [ 3.1415927 ],
+            [ 100 ],
+            [ STDIN ],
+            [ "hello, world!" ]
+        ];
+    }
 }
 
 class EnsureAnyOneOfTest_EnsureNull implements Assurance
 {
-    public function __invoke($item, $fieldOrVarName = "value")
-    {
-        return $this->to($item, $fieldOrVarName);
-    }
+    use InvokeableAssurance;
+    use ListableAssurance;
 
     public function to($item, $fieldOrVarName = "value")
     {
@@ -302,10 +364,8 @@ class EnsureAnyOneOfTest_EnsureNull implements Assurance
 
 class EnsureAnyOneOfTest_EnsureNumeric implements Assurance
 {
-    public function __invoke($item, $fieldOrVarName = "value")
-    {
-        return $this->to($item, $fieldOrVarName);
-    }
+    use InvokeableAssurance;
+    use ListableAssurance;
 
     public function to($item, $fieldOrVarName = "value")
     {
@@ -317,10 +377,8 @@ class EnsureAnyOneOfTest_EnsureNumeric implements Assurance
 
 class EnsureAnyOneOfTest_EnsureString implements Assurance
 {
-    public function __invoke($item, $fieldOrVarName = "value")
-    {
-        return $this->to($item, $fieldOrVarName);
-    }
+    use InvokeableAssurance;
+    use ListableAssurance;
 
     public function to($item, $fieldOrVarName = "value")
     {